+++ libs/langchain-azure-cosmosdb/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./azure_cosmosdb_mongodb.js";
 export * from "./azure_cosmosdb_nosql.js";
-export * from "./caches.js";
 export * from "./chat_histories.js";
+export * from "./caches/caches_nosql.js"
+export * from "./caches/caches_mongodb.js"
