+++ libs/langchain-azure-cosmosdb/src/tests/caches/caches_nosql.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { jest } from "@jest/globals";
 import { FakeEmbeddings, FakeLLM } from "@langchain/core/utils/testing";
-import { AzureCosmosDBNoSQLSemanticCache } from "../index.js";
+import { AzureCosmosDBNoSQLSemanticCache } from "../../index.js";
 
 // Create the mock Cosmos DB client
 const createMockClient = () => {
